+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [5.0.0] - 2023-07-21
 
+### Features
+
+- Changed the Search, Upsert and Compact API from mutable to immutable
+- Upsert now returns the set of new keywords added to the Entry Table
+
 ### Bug Fixes
 
 - add missing `async` keyword for compact callbacks